[ARCH-P4] Move immutable filesystem IO behind adapter boundary
